Modification for the crypto tool - including 2FA for encrypting file

+++ AES_Cryptography_Tool.py
@@ -21,37 +21,59 @@ def encrypt_file():
         print("File not found!")
         return
     
-    key = generate_key()
+    #Prompt the user if they really want to encrypt the file
+    else:
+        prompt = input("Do you really want to encrypt this file (yes/no ?): ")
+        
+        #if no - return
+        if prompt == "no" or prompt == "n":
+            return
+        
+        #if yes - continue
+        elif prompt == "yes" or prompt == "y": 
+            key = generate_key()
+            
+            # Encrypting
+            with open(file_path, "rb") as f:
+                data = f.read()
 
-    # Encrypting
-    with open(file_path, "rb") as f:
-        data = f.read()
-    
-    iv = os.urandom(16)
-    cipher = Cipher(algorithms.AES(key), modes.CBC(iv), backend=default_backend())
-    encryptor = cipher.encryptor()
-    
-    padder = padding.PKCS7(128).padder()
-    padded_data = padder.update(data) + padder.finalize()
-    ciphertext = encryptor.update(padded_data) + encryptor.finalize()
-    
-    encrypted_file_path = file_path + ".enc"
-    with open(encrypted_file_path, "wb") as f:
-        f.write(iv + ciphertext)
-    
-    os.remove(file_path) # Remove the original file path to delete tracks
-    print(f"Your file {file_path} has been encrypted and the original content has been deleted")
-    print(f"Your encrypted file: {encrypted_file_path}\n")
-    print(f"Encryption key (THIS IS IMPORTANT IF YOU WANT TO DECRYPT THE FILE ! SAVE IT): {key.hex()}")
+                iv = os.urandom(16)
+                cipher = Cipher(algorithms.AES(key), modes.CBC(iv), backend=default_backend())
+                encryptor = cipher.encryptor()
+                padder = padding.PKCS7(128).padder()
+                padded_data = padder.update(data) + padder.finalize()
+                ciphertext = encryptor.update(padded_data) + encryptor.finalize()
 
-    # Save the key into a file if users want to
-    save_option = input("Do you want to save the key to a file? (yes/no): ").strip().lower()
-    if save_option == "yes" or save_option == "y":
-        key_file = input("Enter the filename to save the key: ").strip()
-        with open(key_file, "w") as f:
-            f.write(key.hex())
-        print(f"Key saved to {key_file}")
+                encrypted_file_path = file_path + ".enc"
+                with open(encrypted_file_path, "wb") as f:
+                    f.write(iv + ciphertext)
+                
+                os.remove(file_path) # Remove the original file path to delete tracks
+                print(f"Your file {file_path} has been encrypted and the original content has been deleted")
+                print(f"Your encrypted file: {encrypted_file_path}\n")
+                print(f"Encryption key (THIS IS IMPORTANT IF YOU WANT TO DECRYPT THE FILE ! SAVE IT): {key.hex()}")
 
+                # Save the key into a file if users want to
+                save_option = input("Do you want to save the key to a file? (yes/no): ").strip().lower()
+                if save_option == "yes" or save_option == "y":
+                    # Use a loop to make sure the user enter the valid key path
+                    while True:
+                        key_file = input("Enter the filename to save the key: ").strip()
+                        #Check file path valid or not
+                        if ".txt" not in key_file:
+                            print("Invalid key file path !")
+                            continue
+                        else:
+                            with open(key_file, "w") as f:
+                               f.write(key.hex())
+                            print(f"Key saved to {key_file}")
+                            break
+                    
+        #Other inputs - display message Inavlid command and return
+        else:
+            print('Invalid command !')
+            return
+    
 # Function to decrypt the file with the provided key
 def decrypt_file():
     encrypted_file_path = input("Enter the encrypted file that you want to decrypt: ")
